Fix tls-min-version= being ignored (#330)
PeerOptions::parse() called parseOptions() only when 'options='
altered sslOptions instead of delaying the parse to context
creation. This missed the fact that for GnuTLS the tlsMinVersion
was also updating the sslOptions string rather than the
parsedOptions variable later in the configuration process.

Call parseOptions() to reset the parsedOptions value whenever
sslOptions string is altered.